requirements.txt
astroid==2.15.4
certifi==2019.3.9
cffi==1.15.1
chardet==3.0.4
colorama==0.4.6
cryptography==40.0.2
decouple==0.0.7
dill==0.3.6
django-cors-headers==4.0.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-filter==23.2
djangorestframework==3.14.0
flake8==6.1.0
importlib-metadata==6.8.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
pep8==1.7.1
platformdirs==3.5.1
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.3
python-decouple==3.1
pytz==2018.5
sqlparse==0.2.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
urllib3==1.24.2
wincertstore==0.2
wrapt==1.15.0
yapf==0.40.1
zipp==3.16.2